make decisionHosapital to OOP
move decisionHospital method to patient class
