+++ pages/api/patient/add-request-bed.ts
@@ -27,15 +27,9 @@ export default async function addPatient(
         const dateNow = moment(new Date()).format("YYYY-MM-DD HH:mm:SS");
 
         console.log("patientLocation", patientLocation);
-        //TODO
         //1 Query by resource "Bed" > 0
         const availableResource = await database.getResource({ resourceName: "Bed", available: { $gte: 0 }});
-        console.log("availableResource", availableResource);
-
-        const availableHospitalId = availableResource.map((resource) => resource.resourceHospital as string)
-
-        console.log("availableHospitalId", availableHospitalId);
-        
+        const availableHospitalId = availableResource.map((resource) => resource.resourceHospital as string)        
         
         //1.5 query active hospital
         const hospitalData = await database.getHospitals({
@@ -52,36 +46,10 @@ export default async function addPatient(
             }
           ]
         });
-        console.log("hospitalData", hospitalData);
-        //2. calculate distance (google map api)
-        //https://maps.googleapis.com/maps/api/d
-        // destiny : 13.081805, 99.966332|13.081805, 99.966340
-        // origin : 13.081805, 99.966340
-        /**
-         * [
-         *  "13.081805, 99.966332", "13.081805, 99.966332"
-         * ]
-         */
 
-        const hospitalLocation = hospitalData.map((hospital: any) => {
-          return (
-            `${hospital.hospitalLocation?.lat}, ${hospital.hospitalLocation?.long}`
-        )})        
-        console.log("hospitalLocation", hospitalLocation);
-        const destinationLocation = hospitalLocation.join('|');
-        const originLocation = `${patientLocation.lat}, ${patientLocation.long}`
-        console.log("destinationLocation", destinationLocation);
+        //2. calculate distance (google map api)
+        const shortestHospitalIndex = await patient.decisionHospital(hospitalData, patientLocation);
         
-        const result = await axios.get(`${process.env.NEXT_PUBLIC_GOOGLE_API}/json?destinations=${destinationLocation}&origins=${originLocation}&key=${process.env.NEXT_PUBLIC_GOOGLE_KEY}`) as any
-        const distance = result.data.rows[0].elements.map((item: any, i: number) => {
-          return({...item, index: i})
-        })
-        distance.sort((a: any, b: any) => {
-          return(a.distance.value - b.distance.value)
-        }
-        )
-        const shortestHospitalIndex = distance[0].index;
-      
         //3. call addPatient
         const addPatient = await patient.addPatient(
           patientName,
